Renamed all three functions and modified dockerfiles
