+++ lambdas/teamsMakerFront/dockerfile
@@ -1,5 +1,5 @@
 FROM public.ecr.aws/lambda/python:3.11
 
-COPY app.py .
+COPY . ${LAMBDA_TASK_ROOT}
 
 CMD ["app.lambda_handler"]
